% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alval_badges.R
\name{alval_badges}
\alias{alval_badges}
\title{Add Continuous Integration Badges}
\usage{
alval_badges(pkg_path = NULL, gh_username = NULL, travis = TRUE,
  coverage = TRUE, appveyor = TRUE, cran = TRUE)
}
\arguments{
\item{pkg_path}{The path to your package}

\item{gh_username}{Your GitHub username}

\item{travis}{TRUE by default adds .travis.yml to the top level directory and opens travis for configuration}

\item{coverage}{TRUE by default adds coverage reports to a package that is in Travis CI}

\item{appveyor}{TRUE by default adds a basic appveyor.yml to the top level of a package and opens AppVeyor CI.}

\item{cran}{TRUE by default adds cran badge for cran versin or not on cran if not on cran}
}
\value{
Tibble with badges based on user inputs. If not input FALSE will be returned.
}
\description{
Add continuous integration services to the R package. Additionally provide
 badges to your GitHub page to place in the readme.
}
\examples{
\dontrun{
badges = alval_ci(pkg_path = '/Users/alval/Box/Research/aliviateR', gh_username = 'avalcarcel9', coverage = FALSE)
badges$travis_badge
badges$coverage_badge
badges$appveyor_badge}
}
