% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alval_flow.R
\name{alval_flow}
\alias{alval_flow}
\title{Generates an R package}
\usage{
alval_flow(path = NULL, pkg_name = NULL, vignette_name = NULL,
  testing = TRUE, data = TRUE,
  title = "What the Package Does (One Line, Title Case)",
  description = "What the package does (one paragraph).",
  firstname = "Alessandra", lastname = "Valcarcel",
  email = "alval@pennmedicine.com", role = c("aut", "cre"))
}
\arguments{
\item{path}{Path where you will be storing your package}

\item{pkg_name}{Name of package you are creating}

\item{vignette_name}{NULL or name of vignette file to be generated. NULL by default will not generate a vignette or associated files}

\item{testing}{FALSE by default. Logical for building testing material through usethis}

\item{data}{TRUE by default. Logical for adding data to package structure.}

\item{title}{Package title to be put in description}

\item{description}{Description to be put in description}

\item{firstname}{First name. Defaults to my firt name}

\item{lastname}{Last name. Defaults to my last name}

\item{email}{Email. Defaults to my email.}

\item{role}{Default is author and creator}
}
\value{
Folder in specified R path with R package related objects. Vignette may be opened if specified.
}
\description{
Generates an R package according to Ali Valcarcel's workflow. This essentially wraps the
usethis package functions that I always use when creating an R package
}
\examples{
\dontrun{
alval_flow(path = '/Users/alval/Desktop',
pkg_name = "test",
vignette_name = NULL,
testing = FALSE,
title = "What the Package Does (One Line, Title Case)",
description = "What the package does (one paragraph).",
firstname = "Alessandra (Ali)",
middlename = "M.",
lastname = "Valcarcel",
email = "alval@pennmedicine.com",
role = c("aut", "cre"))}
}
