% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{dsc}
\alias{dsc}
\title{Calculate Sørensen's-Dice}
\usage{
dsc(gold_standard, comp_method)
}
\arguments{
\item{gold_standard}{a binary 0,1 vector or nifti object to compare}

\item{comp_method}{a binary 0,1 vector or nifti object to compare}
}
\value{
Returns single Sørensen's-Dice calculation. Should be between 0 and 1.
}
\description{
Calculate Sørensen's-Dice between two vectors or NIFTIs
}
\examples{
\dontrun{
gold_standard = c(1,0,1,0,1,0,0)
comp_method =c(0,0,1,0,1,0,0)
dsc(gold_standard = gold_standard, comp_method = comp_method)}
}
