% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printroxygenheader.R
\name{printroxygenheader}
\alias{printroxygenheader}
\title{Prints Roxygen Header}
\usage{
printroxygenheader(func = TRUE, data = FALSE)
}
\arguments{
\item{func}{TRUE is default to return a generic function header}
}
\value{
Returns a message with header info
}
\description{
Prints a generic Roxygen header. Can be copied and personalized
}
\examples{
\dontrun{printroxygenheader()}

}
